order related functions

+++ sky-server/src/main/java/com/sky/controller/user/OrderController.java
@@ -46,7 +46,7 @@ public class OrderController {
      * @return success Result
      */
     @GetMapping("/reminder/{id}")
-    public Result reminder(@PathVariable("id") Integer id) {
+    public Result reminder(@PathVariable("id") Long id) {
         orderService.reminder(id);
         return Result.success();
     }
